--- a/src/InputHandler.ts
+++ b/src/InputHandler.ts
@@ -21,8 +21,8 @@ import { AttributeData } from 'common/buffer/AttributeData';
 import { IAttributeData, IDisposable } from 'common/Types';
 import { ICoreService, IBufferService, IOptionsService, ILogService, IDirtyRowService } from 'common/services/Services';
 import { ISelectionService } from 'browser/services/Services';
-import { OscHandlerFactory } from 'common/parser/OscParser';
-import { DcsHandlerFactory } from 'common/parser/DcsParser';
+import { OscHandler } from 'common/parser/OscParser';
+import { DcsHandler } from 'common/parser/DcsParser';
 
 /**
  * Map collect to glevel. Used in `selectCharset`.
@@ -241,10 +241,10 @@ export class InputHandler extends Disposable implements IInputHandler {
      * OSC handler
      */
     //   0 - icon name + title
-    this._parser.setOscHandler(0, new OscHandlerFactory((data: string) => this.setTitle(data)));
+    this._parser.setOscHandler(0, new OscHandler((data: string) => this.setTitle(data)));
     //   1 - icon name
     //   2 - title
-    this._parser.setOscHandler(2, new OscHandlerFactory((data: string) => this.setTitle(data)));
+    this._parser.setOscHandler(2, new OscHandler((data: string) => this.setTitle(data)));
     //   3 - set property X in the form "prop=value"
     //   4 - Change Color Number
     //   5 - Change Special Color Number
@@ -505,7 +505,7 @@ export class InputHandler extends Disposable implements IInputHandler {
    * Forward addDcsHandler from parser.
    */
   public addDcsHandler(id: IFunctionIdentifier, callback: (data: string, param: IParams) => boolean): IDisposable {
-    return this._parser.addDcsHandler(id, new DcsHandlerFactory(callback));
+    return this._parser.addDcsHandler(id, new DcsHandler(callback));
   }
 
   /**
@@ -519,7 +519,7 @@ export class InputHandler extends Disposable implements IInputHandler {
    * Forward addOscHandler from parser.
    */
   public addOscHandler(ident: number, callback: (data: string) => boolean): IDisposable {
-    return this._parser.addOscHandler(ident, new OscHandlerFactory(callback));
+    return this._parser.addOscHandler(ident, new OscHandler(callback));
   }
 
   /**