… src/InputHandler.test.ts
… src/TestUtils.test.ts
… src/common/parser/DcsParser.test.ts
… src/common/parser/EscapeSequenceParser.test.ts
… src/common/parser/EscapeSequenceParser.ts
… src/common/parser/OscParser.test.ts
… src/common/parser/Types.d.ts
… src/public/Terminal.ts
… test/api/InputHandler.api.ts
… test/api/Parser.api.ts
… test/benchmark/EscapeSequenceParser.benchmark.ts
… typings/xterm.d.ts