PR to get the parser actions hookable from outside. Shall fix #2332.

TODO:
- [x] review OSC handling and hook
       Also introduces `OscParser` which simplifies `EscapeSequenceParser` and fixes #2342. OSC handlers now can be attached with full state introspection (START, PUT, END) and UTF32 codepoints. The simplified attaching of a single function as handler is now done with a proxy class `OscHandlerFactory`. Payload has a hard limit of 10MB.
- [x] review CSI handling and hook
  Experimental badge removed. Tests were already in place.
- [x] review ESC handling and create public interface
  Interface and tests added.
- [x] review DSC handling and create public interface
  Interface added similar to OSC with callback registration (again via a proxy class to catch the substates). Payload has a hard limit of 10MB.

Also closes #1823.