@jerch 2019-07-26
DCS is halfway done, still missing:
- dcs parser tests
- multiple handler tests
- public interface