@jerch 2019-07-26
@Tyriar Imho there is no formal payload limitation spec'ed for OSC and DCS commands, instead they are meant to work streamlined with chunks. I tried to deal with that by the new `OscParser` / `DcsParser` classes, which provide an interface to the chunk loading to `IDcsHandler` resp. `IOscHandler`.
At this point we clash with our one shot callback handler registration, that gives the callback the payload as one final big chunk, thus we have to store things in between until the sequences gets finished. To not run into OOM I have hardcoded a payload limit for those handlers of 50 MB (done on the proxy classes `OscHandlerFactory` and `DcsHandlerFactory`). Note that most emulators would abort these sequences at much lower limits, maybe @textshell has some numbers here?