@jerch 2019-07-31
Done with the rework of function registration. To see the impact here are some benchmark numbers:
```
"PRINT - a"
  master: 619.79 MB/s
  PR    : 626.24 MB/s
"EXECUTE - \n"
  master: 98.41 MB/s
  PR    : 110.16 MB/s
"ESCAPE - ESC E"
  master: 57.01 MB/s
  PR    : 49.53 MB/s
"ESCAPE with collect - ESC % G"
  master: 22.33 MB/s
  PR    : 62.60 MB/s
"CSI - CSI A"
  master: 92.12 MB/s
  PR    : 86.78 MB/s
"CSI with collect - CSI ? p"
  master: 58.06 MB/s
  PR    : 74.98 MB/s
"CSI with params (short) - CSI 1;2 m"
  master: 101.13 MB/s
  PR    : 97.11 MB/s
"CSI with params (long) - CSI 1;2;3;4;5;6;7;8;9;0 m"
  master: 146.17 MB/s
  PR    : 141.96 MB/s
"OSC (short) - OSC 0;hi ST"
  master: 52.22 MB/s
  PR    : 55.69 MB/s
"OSC (long) - OSC 0;<text> ST"
  master: 48.43 MB/s
  PR    : 65.18 MB/s
"DCS (short)"
  master: 61.88 MB/s
  PR    : 45.79 MB/s
"DCS (long)"
  master: 253.15 MB/s
  PR    : 214.90 MB/s
```
Across all functions its pretty equal. Still interesting that removing the intermediate string concat for "ESCAPE with collect - ESC % G" shows almost 3x boost while "CSI with collect - CSI ? p" is only slightly faster. OSC seems to benefit slightly from the OscParser while DCS decreased (prolly due looping into registered handlers, before it was a single handler).