@jerch 2019-07-31
Introduced at new interface to the API for the callback registration:
```typescript
  export interface IFunctionIdentifier {
    prefix?: string;
    intermediates?: string;
    final: string;
  }
```
This data type is used during callback registration of CSI, DCS and ESC sequences on the parser to create numerical ids as the jump table keys. As shown above this greatly speeds up sequences with intermediates.

Dispatching rules:
- other than before handlers register now on one `IFunctionIdentifier` endpoint, thus if a handler should be reused for a slightly different endpoint, it has to be registered explicitly for that too, example: https://github.com/xtermjs/xterm.js/blob/7252a5921d05ba87de97a36e3d012de6853f0e3c/src/InputHandler.ts#L189-L190
- if no handler was attached a fallback handler will be triggered as before (we currently install those with debug notifications like "unknown sequence XY ...", inspectable with `{logLevel: debug}` in the demo)

Again up for review.