@jerch 2019-08-07
> Any parts in particular that need me to review?

@Tyriar Well the payload limit of 10MB bugs me abit, but I think this can be increased with a string concat patch for chrome once we have a use case for a DCS/OSC command using that much payload.
Also could you have a look on the sub parsers dispose code? Just in case I overlooked something.